+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## Unreleased
+- Adds support for setting range bounds with `new Range(1, 3, "[)")` to set
+  a left-closed, right-open range.
+
 ## 0.1.337 (Oct 9, 2013)
 - First release. Its future is an infinite exclusive range.